@@ -1,12 +1,12 @@
 name: menu_kurukkupattiyal
 description: Swipe(Right to left) the indicator to see the menu, and also can change the position of the indicator.
-version: 0.0.1
+version: 1.0.0
 
 homepage: https://github.com/RalphNoel/menu_kurkkupattiyal
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
