@@ -5,23 +5,23 @@ import 'kurukkupattiyal.dart';
 // This Class is Used to Set Drawer Attributes
 class AlignBuildDrawer extends StatefulWidget {
   const AlignBuildDrawer(
-      {Key key,
-      @required this.listOfContent,
-      @required this.valueListener,
-      @required this.drawerBGColor,
+      {Key? key,
+      required this.listOfContent,
+      required this.valueListener,
+      required this.drawerBGColor,
       this.dividerColor,
       this.dividerSize,
       this.menuIndicatorColor,
       this.onTap})
       : super(key: key);
 
-  final ValueNotifier<double> valueListener;
+  final ValueNotifier<double>? valueListener;
   final List<MenuItem> listOfContent;
   final Color drawerBGColor;
-  final Color dividerColor;
-  final double dividerSize;
-  final Color menuIndicatorColor;
-  final Function(int) onTap;
+  final Color? dividerColor;
+  final double? dividerSize;
+  final Color? menuIndicatorColor;
+  final Function(int)? onTap;
 
   @override
   _AlignBuildDrawerState createState() => _AlignBuildDrawerState();
@@ -33,7 +33,7 @@ class _AlignBuildDrawerState extends State<AlignBuildDrawer> {
     return Align(
       alignment: Alignment(
         1,
-        widget.valueListener.value * 2 - 1,
+        widget.valueListener!.value * 2 - 1,
       ),
       child: Padding(
         padding: const EdgeInsets.only(right: 18.0, top: 10.0, bottom: 10.0),
